Finished backend.py conversion
