+++ spotify2ytmusic/cli.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 
-import json
 import sys
 import os
 import time
@@ -10,15 +9,6 @@ from typing import Optional, Union, Iterator
 
 from . import backend
 from .backend import SongInfo, get_ytmusic
-
-
-def get_playlist_id_by_name(yt: YTMusic, title: str) -> Optional[str]:
-    """Look up a YTMusic playlist ID by name.  Return None if not found."""
-    for pl in yt.get_library_playlists(limit=5000):
-        if pl["title"] == title:
-            return pl["playlistId"]
-
-    return None
 
 
 def list_liked_albums():
@@ -33,7 +23,7 @@ def list_playlists():
     """
     List the playlists on Spotify and YTMusic
     """
-    yt = get_ytmusic()
+    yt = backend.get_ytmusic()
 
     spotify_pls = backend.load_playlists_json()
 
@@ -62,49 +52,6 @@ def create_playlist():
     backend.create_playlist(pl_name)
 
 
-def lookup_song(yt, track_name, artist_name, album_name):
-    """Look up a song on YTMusic
-
-    Given the Spotify track information, it does a lookup for the album by the same
-    artist on YTMusic, then looks at the first 3 hits looking for a track with exactly
-    the same name. In the event that it can't find that exact track, it then does
-    a search of songs for the track name by the same artist and simply returns the
-    first hit.
-
-    The idea is that finding the album and artist and then looking for the exact track
-    match will be more likely to be accurate than searching for the song and artist and
-    relying on the YTMusic algorithm to figure things out, especially for short tracks
-    that might be have many contradictory hits like "Survival by Yes".
-    """
-    albums = yt.search(query=f"{album_name} by {artist_name}", filter="albums")
-    for album in albums[:3]:
-        # print(album)
-        # print(f"ALBUM: {album['browseId']} - {album['title']} - {album['artists'][0]['name']}")
-
-        try:
-            for track in yt.get_album(album["browseId"])["tracks"]:
-                if track["title"] == track_name:
-                    return track
-            # print(f"{track['videoId']} - {track['title']} - {track['artists'][0]['name']}")
-        except Exception as e:
-            print(f"Unable to lookup album ({e}), continuing...")
-
-    songs = yt.search(query=f"{track_name} by {artist_name}", filter="songs")
-    return songs[0]
-
-    #  This would need to do fuzzy matching
-    for song in songs:
-        if (
-            song["title"] == track_name
-            and song["artists"][0]["name"] == artist_name
-            and song["album"]["name"] == album_name
-        ):
-            return song
-        # print(f"SONG: {song['videoId']} - {song['title']} - {song['artists'][0]['name']} - {song['album']['name']}")
-
-    raise ValueError(f"Did not find {track_name} by {artist_name} from {album_name}")
-
-
 def search():
     """Search for a track on ytmusic"""
 
@@ -129,8 +76,8 @@ def search():
 
     args = parse_arguments()
 
-    yt = get_ytmusic()
-    ret = lookup_song(yt, args.track_name, args.artist, args.album)
+    yt = backend.get_ytmusic()
+    ret = backend.lookup_song(yt, args.track_name, args.artist, args.album)
     print(ret)
 
 
@@ -165,7 +112,7 @@ def load_liked_albums():
 
     spotify_pls = backend.load_playlists_json()
 
-    copier(
+    backend.copier(
         backend.iter_spotify_liked_albums(
             spotify_encoding=args.spotify_playlists_encoding
         ),
@@ -203,7 +150,7 @@ def load_liked():
 
     args = parse_arguments()
 
-    copier(
+    backend.copier(
         backend.iter_spotify_playlist(
             None, spotify_encoding=args.spotify_playlists_encoding
         ),
@@ -213,6 +160,7 @@ def load_liked():
     )
 
 
+# @@@@@@@@@@
 def copy_playlist():
     """
     Copy a Spotify playlist to a YTMusic playlist
@@ -250,35 +198,12 @@ def copy_playlist():
         return parser.parse_args()
 
     args = parse_arguments()
-    src_pl_id = args.spotify_playlist_id
-    dst_pl_id = args.ytmusic_playlist_id
-
-    yt = get_ytmusic()
-    if dst_pl_id.startswith("+"):
-        pl_name = dst_pl_id[1:]
-
-        dst_pl_id = get_playlist_id_by_name(yt, pl_name)
-        print(f"Looking up playlist '{pl_name}': id={dst_pl_id}")
-        if dst_pl_id is None:
-            dst_pl_id = backend._ytmusic_create_playlist(
-                yt, title=pl_name, description=pl_name
-            )
-            time.sleep(1)  # seems to be needed to avoid missing playlist ID error
-
-            #  create_playlist returns a dict if there was an error
-            if isinstance(dst_pl_id, dict):
-                print(f"ERROR: Failed to create playlist: {dst_pl_id}")
-                sys.exit(1)
-            print(f"NOTE: Created playlist '{pl_name}' with ID: {dst_pl_id}")
-
-    copier(
-        backend.iter_spotify_playlist(
-            src_pl_id, spotify_encoding=args.spotify_playlists_encoding
-        ),
-        dst_pl_id,
-        args.dry_run,
-        args.track_sleep,
-        yt=yt,
+    backend.copy_playlist(
+        spotify_playlist_id=args.spotify_playlist_id,
+        ytmusic_playlist_id=args.ytmusic_playlist_id,
+        track_sleep=args.track_sleep,
+        dry_run=args.dry_run,
+        spotify_playlists_encoding=args.spotify_playlists_encoding,
     )
 
 
@@ -286,7 +211,6 @@ def copy_all_playlists():
     """
     Copy all Spotify playlists (except Liked Songs) to YTMusic playlists
     """
-    yt = get_ytmusic()
 
     def parse_arguments():
         parser = ArgumentParser()
@@ -310,118 +234,8 @@ def copy_all_playlists():
         return parser.parse_args()
 
     args = parse_arguments()
-    spotify_pls = backend.load_playlists_json()
-
-    for src_pl in spotify_pls["playlists"]:
-        if str(src_pl.get("name")) == "Liked Songs":
-            continue
-
-        pl_name = src_pl["name"]
-        if pl_name == "":
-            pl_name = f"Unnamed Spotify Playlist {src_pl['id']}"
-
-        dst_pl_id = get_playlist_id_by_name(yt, pl_name)
-        print(f"Looking up playlist '{pl_name}': id={dst_pl_id}")
-        if dst_pl_id is None:
-            dst_pl_id = backend._ytmusic_create_playlist(
-                yt, title=pl_name, description=pl_name
-            )
-            time.sleep(1)  # seems to be needed to avoid missing playlist ID error
-
-            #  create_playlist returns a dict if there was an error
-            if isinstance(dst_pl_id, dict):
-                print(f"ERROR: Failed to create playlist: {dst_pl_id}")
-                sys.exit(1)
-            print(f"NOTE: Created playlist '{pl_name}' with ID: {dst_pl_id}")
-
-        copier(
-            backend.iter_spotify_playlist(
-                src_pl["id"], spotify_encoding=args.spotify_playlists_encoding
-            ),
-            dst_pl_id,
-            args.dry_run,
-            args.track_sleep,
-        )
-        print("\nPlaylist done!\n")
-
-    print("All done!")
-
-
-def copier(
-    src_tracks: Iterator[SongInfo],
-    dst_pl_id: Optional[str] = None,
-    dry_run: bool = False,
-    track_sleep: float = 0.1,
-    *,
-    yt: Optional[YTMusic] = None,
-):
-    if yt is None:
-        yt = get_ytmusic()
-
-    if dst_pl_id is not None:
-        try:
-            yt_pl = yt.get_playlist(playlistId=dst_pl_id)
-        except Exception as e:
-            print(f"ERROR: Unable to find YTMusic playlist {dst_pl_id}: {e}")
-            print(
-                "       Make sure the YTMusic playlist ID is correct, it should be something like "
-            )
-            print("      'PL_DhcdsaJ7echjfdsaJFhdsWUd73HJFca'")
-            sys.exit(1)
-        print(f"== Youtube Playlist: {yt_pl['title']}")
-
-    tracks_added_set = set()
-    duplicate_count = 0
-    error_count = 0
-
-    for src_track in src_tracks:
-        print(f"Spotify:   {src_track.title} - {src_track.artist} - {src_track.album}")
-
-        try:
-            dst_track = lookup_song(
-                yt, src_track.title, src_track.artist, src_track.album
-            )
-        except Exception as e:
-            print(f"ERROR: Unable to look up song on YTMusic: {e}")
-            error_count += 1
-            continue
-
-        yt_artist_name = "<Unknown>"
-        if "artists" in dst_track and len(dst_track["artists"]) > 0:
-            yt_artist_name = dst_track["artists"][0]["name"]
-        print(
-            f"  Youtube: {dst_track['title']} - {yt_artist_name} - {dst_track['album']}"
-        )
-
-        if dst_track["videoId"] in tracks_added_set:
-            print("(DUPLICATE, this track has already been added)")
-            duplicate_count += 1
-        tracks_added_set.add(dst_track["videoId"])
-
-        if not dry_run:
-            exception_sleep = 5
-            for _ in range(10):
-                try:
-                    if dst_pl_id is not None:
-                        yt.add_playlist_items(
-                            playlistId=dst_pl_id,
-                            videoIds=[dst_track["videoId"]],
-                            duplicates=False,
-                        )
-                    else:
-                        yt.rate_song(dst_track["videoId"], "LIKE")
-                    break
-                except Exception as e:
-                    print(
-                        f"ERROR: (Retrying add_playlist_items: {dst_pl_id} {dst_track['videoId']}) {e} in {exception_sleep} seconds"
-                    )
-                    time.sleep(exception_sleep)
-                    exception_sleep *= 2
-
-        if track_sleep:
-            time.sleep(track_sleep)
-
-    print()
-    print(
-        f"Added {len(tracks_added_set)} tracks, encountered {duplicate_count} duplicates, {error_count} errors"
+    backend.copy_all_playlists(
+        track_sleep=args.track_sleep,
+        dry_run=args.dry_run,
+        spotify_playlists_encoding=args.spotify_playlists_encoding,
     )
